Improves tests
